third_party/WebKit/LayoutTests/fast/css/pseudo-in-range-invalid-value-expected.txt

CONSOLE WARNING: line 5: The specified value "ABC" is not a valid number. The value must match to the following regular expression: -?(\d+|\d+\.\d+|\.\d+)([eE][-+]?\d+)?
Tests that invalid values are considered in-range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector("input[type=number]:in-range").id is "number1"
PASS document.querySelector("input[type=range]:in-range").id is "range1"
PASS document.querySelectorAll(":in-range").length is 2
PASS successfullyParsed is true

TEST COMPLETE

 
